feat: Add option to view  movies by trending daily
Allow user to view movies by selecting the trending daily option menu item.
Create and an extra option menu item in the MainActivity. Make modification
to NetworkUtils to accomodate for the different url and query parameters
for trending movies.

This feature allows the user to have a view a list that is more dynamic
(in terms of updating movies) than the PopularMovies list (where top
movies can remain in the list for quite some time)