feat: Introduce bottom navigation bar
Add bottom navigation bar for navigation between movie sources.
Also add feature to bring up soft keyboard when on search screen without having
to touch the editText first.

These changes make it easier for the user to get a lot of the app
functionality without having to stretch all the way to the top corner
of the screen.

This commit does not yet remove the movie source options from the options menu.
The options co-exist with each other (in the options menu and bottom nav bar).
Items in the options menu behave as expected when the bottom Nav bar is used
to navigate. However, items in the bottom nav bar do not respond when items
in the options menu are used. This is intentional. As it is expected that the
source options in the options menu will be removed.